BE: (#0004) database package renamed to db
